requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy[asyncio]
aiohttp
pydantic
python-dotenv
